(API Change) Expand ClientId to a proper PCB-like struct.
Rationale: We want clients to be able to choose between different crypto
implementations (pure software crypto with storage in the filesystem
vs. hardware-accelerated crypto with tamperproof storage in a secure
element), and we need to store this choice inside Trussed. So we have to
change ClientId from a simple typedef to littlefs2::PathBuf to a proper
struct that's starting to resemble a PCB.

In this intial implementation, the choice has to be specified on endpoint
creation. While this could easily be made more flexible (e.g. by introducing
an additional syscall), the calling app would probably require a reset when
switching the provider, as keys are usually not transferrable out of
the secure element.