Add a miscellaneous GUI control syscall.
This command adds an extensible GUIControl syscall with enum-based Command
and Response types as a catch-all for future platform-specific GUI actions.

The first GUI prototype has a USB-C plug and thus needs support for rotating
the view, so graphics are readable regardless of whether the device is
plugged in from the left or the right. This necessitates the first
Command/Response pair for GUIControl: SetOrientation.